pyproject.toml
[project]
name = "query-to-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-search-results>=2.4.2",
    "openai>=1.79.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
]
